FIX-1013: move compress_store_swizzle_mask_num
