+++ include/eve/detail/function/simd/common/compress_using_masks.hpp
@@ -9,7 +9,7 @@
 
 #include <eve/traits/max_scalar_size.hpp>
 
-#include <eve/detail/function/compress_store_swizzle_mask_num.hpp>
+#include <eve/detail/compress/mask_num.hpp>
 #include <eve/detail/function/byte_16_runtime_shuffle.hpp>
 
 #include <eve/function/count_true.hpp>
