+++ test/unit/internals/CMakeLists.txt
@@ -14,7 +14,7 @@ add_custom_target(unit.internals.exe )
 ##==================================================================================================
 make_unit( "unit.internals" aggregation.cpp                     )
 make_unit( "unit.internals" byte_16_runtime_shuffle.cpp         )
-make_unit( "unit.internals" compress_store_swizzle_mask_num.cpp )
+make_unit( "unit.internals" compress/mask_num.cpp               )
 make_unit( "unit.internals" dispatch.cpp                        )
 make_unit( "unit.internals" horn.cpp                            )
 make_unit( "unit.internals" horn1.cpp                           )
